[NEAR-110] Refactor: 라우터 분리
- 스웨거 문서상 분리된 태그로 관리하기 위함
