+++ app/domains/streams/admin/stream_router.py
@@ -2,12 +2,10 @@ from fastapi import (
     APIRouter,
     Body,
     Depends,
-    File,
     Path,
     Query,
     Request,
     Response,
-    UploadFile,
     status,
 )
 from fastapi.responses import HTMLResponse
@@ -15,9 +13,6 @@ from fastapi.templating import Jinja2Templates
 
 from app.domains.streams import deps as streams_deps
 from app.domains.streams.admin.schemas import (
-    ConcertCreateRequest,
-    ConcertDetailResponse,
-    ConcertResponse,
     SessionCreateRequest,
     SessionResponse,
     SessionUpdateRequest,
@@ -25,104 +20,10 @@ from app.domains.streams.admin.schemas import (
 )
 from app.domains.streams.admin.service import StreamAdminService
 from app.domains.streams.deps import get_admin_user
-from app.domains.streams.models import Concert
 from app.domains.users.models import User
 
-router = APIRouter(prefix="/admin/streams", tags=["[Admin] 스트리밍 관리"])
+router = APIRouter(prefix="/admin/streams", tags=["스트리밍 관리"])
 templates = Jinja2Templates(directory="templates")
-
-
-@router.post(
-    "/concerts",
-    status_code=status.HTTP_201_CREATED,
-    response_model=ConcertResponse,
-    summary="콘서트 메타정보 생성",
-    description="신규 콘서트의 기본 메타데이터(제목, 카테고리 등)를 생성합니다. \
-    인프라 리소스는 생성되지 않습니다.",
-)
-async def create_concert(
-    data: ConcertCreateRequest,
-    current_admin: User = Depends(get_admin_user),
-    service: StreamAdminService = Depends(streams_deps.get_stream_admin_service),
-) -> Concert:
-    return await service.create_concert(data, current_admin)
-
-
-@router.get(
-    "/concerts",
-    response_model=list[ConcertResponse],
-    summary="콘서트 목록 조회",
-)
-async def list_concerts(
-    response: Response,
-    cursor: int | None = Query(None, description="마지막으로 조회된 콘서트 ID"),
-    limit: int = Query(20, ge=1, le=100),
-    current_admin: User = Depends(get_admin_user),
-    service: StreamAdminService = Depends(streams_deps.get_stream_admin_service),
-) -> list[Concert]:
-    concerts, next_cursor = await service.list_concerts(current_admin, limit, cursor)
-
-    if next_cursor is not None:
-        response.headers["X-Next-Cursor"] = str(next_cursor)
-    return concerts
-
-
-@router.get(
-    "/concerts/{concert_id}",
-    response_model=ConcertDetailResponse,
-    summary="콘서트 상세 조회",
-)
-async def get_concert_detail(
-    concert_id: int = Path(..., description="콘서트 ID"),
-    current_admin: User = Depends(get_admin_user),
-    service: StreamAdminService = Depends(streams_deps.get_stream_admin_service),
-) -> ConcertDetailResponse:
-    return await service.get_concert_detail(concert_id, current_admin)
-
-
-@router.patch(
-    "/concerts/{concert_id}",
-    response_model=ConcertResponse,
-    summary="콘서트 정보 수정",
-)
-async def update_concert(
-    concert_id: int = Path(...),
-    data: ConcertCreateRequest = Body(...),
-    current_admin: User = Depends(get_admin_user),
-    service: StreamAdminService = Depends(streams_deps.get_stream_admin_service),
-) -> Concert:
-    return await service.update_concert(concert_id, data, current_admin)
-
-
-@router.delete(
-    "/concerts/{concert_id}",
-    status_code=status.HTTP_204_NO_CONTENT,
-    summary="콘서트 및 관련 인프라 전체 삭제",
-)
-async def delete_concert(
-    concert_id: int = Path(...),
-    current_admin: User = Depends(get_admin_user),
-    service: StreamAdminService = Depends(streams_deps.get_stream_admin_service),
-) -> None:
-    return await service.delete_concert_with_infrastructure(concert_id, current_admin)
-
-
-@router.patch(
-    "/concerts/{concert_id}/thumbnail",
-    response_model=ConcertResponse,
-    summary="콘서트 썸네일 이미지 생성 및 수정",
-    description="콘서트 썸네일 이미지를 업로드 또는 수정합니다.",
-)
-async def update_concert_thumbnail(
-    current_admin: User = Depends(get_admin_user),
-    concert_id: int = Path(..., description="콘서트 ID"),
-    thumbnail_file: UploadFile = File(..., description="썸네일 이미지 파일"),
-    service: StreamAdminService = Depends(streams_deps.get_stream_admin_service),
-) -> Concert:
-    # 인증된 current_admin 객체를 서비스로 직접 전달
-    return await service.update_concert_thumbnail(
-        concert_id=concert_id, file=thumbnail_file, user=current_admin
-    )
 
 
 @router.post(
